requirements.txt
python-telegram-bot[job-queue]==21.10
Pillow==11.1.0